+++ test/it/parser/src/main/java/org/apache/shardingsphere/test/it/sql/parser/internal/cases/parser/jaxb/segment/impl/expr/ExpectedExpression.java
@@ -110,6 +110,9 @@ public final class ExpectedExpression extends AbstractExpectedSQLSegment {
     @XmlElement(name = "outer-join-expression")
     private ExpectedOuterJoinExpression outerJoinExpression;
     
+    @XmlElement(name = "interval-expression-projection")
+    private ExpectedIntervalExpressionProjection intervalExpressionProjection;
+    
     @XmlElement(name = "interval-expression")
     private ExpectedIntervalExpression intervalExpression;